Refactor commands to individual packages.
First step towards cleaning up the command code & duplication.
